Fix captcha.php

+++ controller/captcha.php
@@ -50,7 +50,7 @@ function generateCaptcha(){
         imageline($im, $rand1, $rand2, $rand1, $rand2, $net);
     }
 
-    $x = rand(5, $width/(7/2));
+    $x = rand(5, (int)($width/(7/2)));
 
     imagerectangle($im, 0, 0, $width-1, $height-1, $frame);
